fix: get tests back to work by fixing broken imports
also: incorporate new test as suggested by @tmpolaczyk
