+++ validations/src/validations.rs
@@ -2313,29 +2313,4 @@ mod tests {
         assert_eq!(reward(3_500_000 * 65), 0);
         assert_eq!(reward(3_500_000 * 100), 0);
     }
-
-    #[test]
-    fn test_counter() {
-        let mut counter = Counter::new(7);
-        counter.increment(6);
-        assert_eq!(counter.max_val, 1);
-        assert_eq!(counter.max_pos, Some(6));
-
-        counter.increment(6);
-        assert_eq!(counter.max_val, 2);
-        assert_eq!(counter.max_pos, Some(6));
-
-        counter.increment(0);
-        assert_eq!(counter.max_val, 2);
-        assert_eq!(counter.max_pos, Some(6));
-
-        counter.increment(0);
-        counter.increment(0);
-        assert_eq!(counter.max_val, 3);
-        assert_eq!(counter.max_pos, Some(0));
-
-        counter.increment(6);
-        assert_eq!(counter.max_val, 3);
-        assert_eq!(counter.max_pos, None);
-    }
 }
